Various changes working toward a better working index
Right now I think the problem is neighbors aren't being collected right.
Not enough edges are being made or saved, and everything starts at node 0
of the layer so the entry points are all skewed. We should maybe be
walking from our best neighor down to the neighbor's best neighbor in
the LAYER DOWN. This might be the issue is indexing and searching down
the layers.