Rewrote a bunch of the core algorithms, but they're still busted
I'm tired. I think this is a graph traversal problem, but figuring out
where to start in the graph when looking for nearest neighbors and how
far to fan out while looking downwards is hard. It's easy to do it
exhaustive, but it needs to be smarter than that so it doesn't fan out
to everything.