A bunch of hacking and logging chasing down node connections
I found a case where edges aren't fully reciprocal. One vertex may have
an edge to another, but the peer drops the incoming edge because it's
inferior to what it already had relative to itself.

Somehow this leads to best data being unreachable from the entry-point
because there are enough nodes in a cluster that connections to further
out data seem to no longer exist.