Higher layer and probabilty changes the shape of index and helps searching
But it seems to take longer. I still think there's a problem where nodes
inserted lower need to look up and link to their higher level peers to
create relevant connections later to avoid order-of-insert problems.
Nearby data may not be reachable in a meaningful way from higher levels
if they are not inserted in "lucky" order.