WIP Some changes in structure to fork/join but still not correct
I think the problem is that the graph searching is now directionless.
It just starts following edges, adding things into the better set if
possible, but if it starts going the wrong way initially (because
anything is better than nothing) then it can end up in a weird corner
with no way (or need, since full enough) to go back far enough to
try a better path.