WIP using a "channel" to burn down candidates
The problem with this is that the stopping condition is not clear. Or
at least I'm having trouble coding it. It's something like "when there's
no candidates left then stop" but you have a bunch of things waiting,
and one thing searching existing sets. Maybe it needs some kind of
counter? A countdown latch? A semaphore? idk :(

Maybe it just needs to go back to fork/join tasks and have each one
aggressively while looping to burn down the list. Then they can just
naturally exit as it sees fit without a need to track the state of all
workers.